Fix some of the tests
